refactoring test methods to just be in __init__.py
